packages/backend/scraper/requirements.txt
beautifulsoup4
pandas
selenium
mysql-connector-python
tqdm
requests
dotenv